pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ansys-stk-core"
dynamic = ["version"]
description = "A Python API for STK"
readme = "README.rst"
requires-python = ">=3.6"
license = {file = "LICENSE"}
authors = [
    {name = "ANSYS, Inc.", email = "pyansys.support@ansys.com"},
]
maintainers = [
    {name = "PyAnsys developers", email = "pyansys.maintainers@ansys.com"},
]

classifiers = [
    # TODO: remove all 'Private' classifiers prior to the public release
    "Private :: Do Not Upload",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
]

[project.optional-dependencies]
tests = [
    "pytest==7.4.0",
    "pytest-cov==4.1.0",
    "pytz==2023.3",
]
doc = [
    "ansys-sphinx-theme==0.10.3",
    "enum-tools[sphinx]==0.10.0",
    "numpydoc==1.5.0",
    "sphinx==7.1.2",
    "sphinx-design==0.5.0",
    "sphinx-jinja==2.0.2",
    "sphinx-copybutton==0.5.2",
]

[tool.flit.module]
name = "ansys.stk.core"

[project.urls]
Homepage = "https://www.ansys.com/products/missions/ansys-stk"
Documentation = "https://stk.docs.pyansys.com/"
Source = "https://github.com/pyansys/pystk"
Tracker = "https://github.com/pyansys/pystk/issues"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
force_sort_within_sections = true
line_length = 80
default_section = "THIRDPARTY"
src_paths = ["examples", "doc", "src", "tests"]

[tool.coverage.run]
relative_files = true

[tool.coverage.paths]
source = [
    "src",
    "ansys/stk",
]

[tool.coverage.report]
show_missing = true
